requirements.txt
codecarbon==2.4.2
matplotlib==3.9.0
pycryptodome==3.20.0
twofish==0.3.0
xtea==0.7.1
